requirements.txt
django-rest-knox==4.2.0
djangorestframework==3.14.0
django-allauth==0.60.1
django_rest_passwordreset==1.3.0
psycopg2-binary==2.9.9